add bindings, plugins, and view models modules, add hard-coded arches module, add the ability to download and inject widget templates and widget components, this would all have to be refactored.
I would break out the downloading of templates and components into it's own method, separate from injecting them into the DOM.
Then the injection of templates and components should happen based on the files stored in the file system.

Additionally, the file download plugin may not be the best, and more investigation should be put there.